Update 18/04/2024
- Sensor Node changed to work with the Arduino Nano ESP32
- Change readings to int16_t to be supported across all boards
- The node-red tree now displays the name of each node on top of its id
- The code of the node-red was simplified

+++ main_node/main_node.h
@@ -13,8 +13,8 @@ const int NAME_LENGTH = 7;
 
 struct Sensor_Node { // os inteiros têm tamanhos diferentes num e noutro
   char name[NAME_LENGTH];
-  float temperature;
-  float phototransistor;
+  int16_t temperature;
+  int16_t phototransistor;
 };
 
 struct Student_Node {
